Add AppIdentifier to Issuer type
