+++ Tests/SnapshotTests/Components/UPIComponentUITests.swift
@@ -14,8 +14,8 @@ class UPIComponentUITests: XCTestCase {
     // TODO: - Revert this change once the UPI apps response is not mocked.
     private lazy var paymentMethod: UPIPaymentMethod = .init(
         type: .upi,
-        name: "upi"
-//        apps: upiApps
+        name: "upi",
+        apps: upiApps
     )
     private var context: AdyenContext {
         Dummy.context
@@ -26,9 +26,9 @@ class UPIComponentUITests: XCTestCase {
     }
 
     private var upiApps: [Issuer] = [
-        Issuer(identifier: "bhim", name: "BHIM"),
-        Issuer(identifier: "gpay", name: "Google Pay"),
-        Issuer(identifier: "phonepe", name: "PhonePe")
+        Issuer(identifier: "bhim", name: "BHIM", appIdentifier: .init(scheme: "bhim")),
+        Issuer(identifier: "gpay", name: "Google Pay", appIdentifier: .init(scheme: "tez")),
+        Issuer(identifier: "phonepe", name: "PhonePe", appIdentifier: .init(scheme: "phonepe"))
     ]
 
     override func setUpWithError() throws {
